Rename activate to connect
